requirements.txt
openai
spacy
python-dotenv

